pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-evals"
version = "0.1.0"
description = "Product-minded LLM eval framework with deterministic checks, persona scoring, judge-based evaluation, reporting, and CI regression gates"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Vignesh" }]
dependencies = [
    "pydantic>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "anthropic>=0.40",
    "openai>=1.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
llm-evals = "llm_evals.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_evals"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
